backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
python-dotenv
psycopg2-binary
alembic
httpx
pytest
aiofiles
Pillow
redis
fastapi-cache2
rate-limit
pytest-cov
pytest-asyncio
locust